Move some logic from file_*.rs to bmbt_rec.rs
