+++ src/libxfuse/file_extent_list.rs
@@ -51,58 +51,7 @@ impl<R: BufRead + Reader + Seek> File<R> for FileExtentList {
     }
 
     fn lseek(&self, _buf_reader: &mut R, offset: u64, whence: i32) -> Result<u64, i32> {
-        let sb = SUPERBLOCK.get().unwrap();
-
-        let dblock = offset >> sb.sb_blocklog;
-        match self.bmx.0.partition_point(|entry| entry.br_startoff <= dblock) {
-            0 => {
-                // A hole at the beginning of the file
-                if whence == libc::SEEK_HOLE {
-                    Ok(offset)
-                } else {
-                    self.bmx.first()
-                        .map(|b| b.br_startoff << sb.sb_blocklog)
-                        .ok_or(libc::ENXIO)
-                }
-            },
-            i => {
-                let cur_entry = &self.bmx.0[i - 1];
-                let br_end = cur_entry.br_startoff + cur_entry.br_blockcount;
-                if dblock < br_end {
-                    // In a data region
-                    if whence == libc::SEEK_HOLE {
-                        // Scan for the next hole
-                        for j in (i - 1)..self.bmx.0.len() - 1 {
-                            let before = &self.bmx.0[j];
-                            let after = &self.bmx.0[j + 1];
-                            let br_end = before.br_startoff + before.br_blockcount;
-                            if after.br_startoff > br_end {
-                                return Ok(br_end << sb.sb_blocklog);
-                            }
-                        }
-                        // Reached EOF without finding another hole.  Return the virtual hole at
-                        // EOF
-                        let entry = self.bmx.0.last().unwrap();
-                        let br_end = entry.br_startoff + entry.br_blockcount;
-                        Ok(br_end << sb.sb_blocklog)
-                    } else {
-                        Ok(offset)
-                    }
-                } else {
-                    // In a hole
-                    if whence == libc::SEEK_HOLE {
-                        Ok(offset)
-                    } else {
-                        match self.bmx.0.get(i) {
-                            Some(next_entry) => {
-                                Ok(next_entry.br_startoff << sb.sb_blocklog)
-                            },
-                            None => Err(libc::ENXIO)
-                        }
-                    }
-                }
-            }
-        }
+        self.bmx.lseek(offset, whence)
     }
 
     fn size(&self) -> XfsFsize {
